requirements.txt
openrgb-python>=0.3.0
psutil>=5.9.0
